root/requirements.txt
flask
celery[redis]
flask-sqlalchemy
flask-restful
flask-security-too
flask-caching
bcrypt
httplib2
redis